core/standards/web-apps/media/semantics/seeking/seek_to_unbuffered.html

<!doctype html>
<title>HTMLMediaElement.buffered: Seeking to an unbuffered part of already loading media - webm</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/core/standards/web-apps/media/support/common.js"></script>
<p id="log">FAIL (script didn't run)</p>
<audio autoplay controls></audio>
<video autoplay controls></video>
<script>
var tests = init_tests("HTMLAudioElement.buffered: Seeking to an unbuffered part of already loading audio - webm", "HTMLVideoElement.buffered: Seeking to an unbuffered part of already loading video - webm", {timeout:10000});
tests.forEach(function(vars) {
    var t = vars[0];
    t.step(function() {
    var tag_name = vars[1];
    var tolerance = 10;
    var node = document.getElementsByTagName(tag_name)[0];
    node.addEventListener('loadedmetadata', startTest, false);
    node.src = '/core/standards/web-apps/media/network/range-request-log/range-request.php?rate=20000&fileloc=../../support/preload.webm&nocache=' + Math.random();

    function startTest() {
        node.addEventListener('seeked', endTest, false);
        node.currentTime = 50;
    }

    function endTest() {
        t.step(function() {
        opera.postError(node.buffered.end(1));
        assert_true(Math.abs(node.buffered.end(1) - 50) < tolerance, "'buffered.end(1)' should be 50 \u00B1 " + tolerance);
        t.done();
        });
    }
    });
});
</script>
